+++ test/test.js
@@ -1,7 +1,13 @@
 
 var assert = require('assert')
+var crypto = require('crypto')
+var proxyquire = require('proxyquire')
 
-var uid = require('..')
+var uid = proxyquire('..', {
+  crypto: {
+    randomBytes: randomBytes
+  }
+})
 
 describe('uid-url', function () {
   describe('uid()', function () {
@@ -28,6 +34,24 @@ describe('uid-url', function () {
         done()
       })
     })
+
+    describe('when PRNG not seeded', function () {
+      before(function () {
+        randomBytes.seeded = false
+      })
+
+      after(function () {
+        randomBytes.seeded = true
+      })
+
+      it('should still generate uid', function (done) {
+        uid(18, function (err, val) {
+          if (err) return done(err)
+          assert.equal(24, Buffer.byteLength(val))
+          done()
+        })
+      })
+    })
   })
 
   describe('uid.sync()', function () {
@@ -42,5 +66,35 @@ describe('uid-url', function () {
       assert(!~val.indexOf('/'))
       assert(!~val.indexOf('='))
     })
+
+    describe('when PRNG not seeded', function () {
+      before(function () {
+        randomBytes.seeded = false
+      })
+
+      after(function () {
+        randomBytes.seeded = true
+      })
+
+      it('should still generate uid', function () {
+        var val = uid.sync(18)
+        assert.equal(24, Buffer.byteLength(val))
+      })
+    })
   })
 })
+
+function randomBytes(length, callback) {
+  if (randomBytes.seeded !== false) {
+    return crypto.randomBytes(length, callback)
+  }
+
+  // The crazy not seeded error
+  var err = new Error('error:24064064:random number generator:SSLEAY_RAND_BYTES:PRNG not seeded')
+
+  if (!callback) {
+    throw err
+  }
+
+  callback(err)
+}